code/samg/requirements.txt
numba==0.59.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
